Move models into their own separate modules
This shows a clear distinction between the two sets of models and helps
clarify the design.
