+++ koiki/woocommerce/models.py
@@ -1,5 +1,13 @@
-from koiki.vendor import Vendor
+class Vendor():
+    def __init__(self, id, name):
+        self.id = id
+        self.name = name
 
+    def __eq__(self, other):
+        if not isinstance(other, Vendor):
+            return NotImplemented
+
+        return self.id == other.id and self.name == other.name
 
 class LineItem():
     def __init__(self, line_item):
